+++ CHANGELOG.md
@@ -5,10 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v4.5.0
+
+* Added more flags to configure flushing intervals.
+
+  Available flags are `batch-block-flush-interval`, `batch-row-flush-interval` and `live-block-flush-interval`, check `substreams-sink-sql run --help` for full documentation about the flags.
+
 ## v4.4.0
 
 * Added support for the Clickhouse `Date` type.
-* Added more flags to configure flushing intervals. Available flags are `batch-block-flush-interval`, `batch-row-flush-interval` and `live-block-flush-interval`.
 * Deprecated the existing `flush-interval` flag in favor of `batch-block-flush-interval`.
 
 * Fixed handling of the Clickhouse `Array` type.